pyproject.toml
[project]
name = "pink-flamingos-elt-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "google-auth>=2.43.0",
    "google-cloud-storage>=3.5.0",
    "minio>=7.2.18",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.4",
]
